seprated tests into different folders
